requirements.txt
absl-py
aiohttp
aiosignal
amqp
annotated-types
anyio
APScheduler
asgiref
asttokens
astunparse
attrs
billiard
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
comm
contourpy
coverage
cryptography
cycler
dataclasses-json
debugpy
decorator
defusedxml
distlib
distro
dj-database-url
Django
django
django-autocomplete-light
django-cors-headers
django-countries
django-filter
django-phonenumber-field
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djhacker
djoser
executing
filelock
flatbuffers
fonttools
frozenlist
gast
google-pasta
greenlet
grpcio
gunicorn
h11
h5py
httpcore
httpx
idna
iniconfig
ipykernel
ipython
jedi
joblib
jsonpatch
jsonpointer
jupyter_client
jupyter_core
keras
kiwisolver
kombu
Markdown
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
ml-dtypes
multidict
mypy-extensions
namex
nest-asyncio
numpy
oauthlib
openai==0.28.0
opt-einsum
optree
orjson
packaging
pandas
parso
phonenumbers
pillow
platformdirs
pluggy
prompt_toolkit
protobuf
psutil
psycopg2
psycopg2-binary
pure-eval
pycparser
pydantic
pydantic_core
Pygments
PyJWT
pyparsing
PyPDF2
pytest
pytest-cov
pytest-django
python-dateutil==2.9.0.post0
python-dotenv
python3-openid
pytz==2024.1
PyYAML
pyzmq
requests
requests-oauthlib
rich
scikit-learn
scipy
seaborn
setuptools
six
sniffio
social-auth-app-django
social-auth-core
SQLAlchemy
sqlparse
stack-data
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-intel
termcolor
threadpoolctl
tornado
tqdm
traitlets
typing-inspect
typing_extensions
tzdata
tzlocal
urllib3
vine
virtualenv
wcwidth
Werkzeug
wheel
wrapt
yarl
